backend/AuthService/pyproject.toml
[tool.poetry]
name = "hospital"
version = "0.1.0"
description = ""
authors = ["Rubelewaskes <131644269+Rubelewaskes@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = {extras = ["all"], version = "^0.115.5"}
pyjwt = {extras = ["crypto"], version = "^2.10.0"}
pydantic = "^2.10.1"
python-multipart = "^0.0.17"
pathlib = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
